[TASK] removed different stuff, added filtering options
