--- Resources/Private/Language/de.locallang.xlf
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8" standalone="yes" ?>
-<xliff version="1.0">
-	<file source-language="en" target-language="de" datatype="plaintext" original="messages" date="2014-03-18T11:48:24Z" product-name="pxa_dealers">
-		<header/>
-		<body>
-
-			<trans-unit id="error.NotNumeric">
-				<target>Zipcode should be numeric</target>
-			</trans-unit>
-			
-			<trans-unit id="placeholder">
-				<target>PLZ oder Stadt...</target>
-			</trans-unit>
-			<trans-unit id="submitvalue">
-				<target>Finden</target>
-			</trans-unit>
-			<trans-unit id="searchDidntStart">
-				<target>Bitte geben Sie eine PLZ oder Stadt ein um die Suche zu starten</target>
-			</trans-unit>	
-			
-			<trans-unit id="filtrering">
-				<target>Filtrering</target>
-			</trans-unit>
-
-			<trans-unit id="error.EmptyValue">
-				<target>Geben Sie eine PLZ oder eine Stadt</target>
-			</trans-unit>
-
-			<trans-unit id="country_list_city_zip_search">
-				<source>search for city or zip...</source>
-				<target>PLZ oder Ort angeben</target>
-			</trans-unit>
-
-			<trans-unit id="dealers_found">
-                <source>dealers found</source>
-                <target>gefundene Shops</target>
-            </trans-unit>
-
-			<trans-unit id="error.NoResult">
-				<target>Keinen Händler gefunden. Suchen Sie erneut mit einer anderen PLZ oder einem Ort.</target>
-			</trans-unit>
-
-			<trans-unit id="notDealersFoundFilteringMessage">
-				<target>Keine Shops gefunden. Starten Sie eine neue Suche mit einer anderen PLZ oder entfernen Sie die Filteroptionen.</target>
-			</trans-unit>
-
-            <trans-unit id="categoriesFilter.radio.all">
-                <source>All</source>
-                <target>Alle Shops</target>
-            </trans-unit>
-
-			<trans-unit id="country_list_select_country">
-				<source>Select country</source>
-				<target>Land auswählen >></target>
-			</trans-unit>
-
-			<trans-unit id="country_list.all_countries">
-				<source>All countries</source>
-				<target>Alle Länder</target>
-			</trans-unit>
-
-		</body>
-	</file>
-</xliff>
