+++ Resources/Private/Language/locallang.xlf
@@ -8,6 +8,13 @@
 				<source>Show on map</source>
 			</trans-unit>
 
+            <trans-unit id="fe.placeholder">
+                <source>Enter country, postcode or city</source>
+            </trans-unit>
+            <trans-unit id="fe.submitvalue">
+                <source>Search</source>
+            </trans-unit>
+
             <trans-unit id="js.infoWindowPhone">
                 <source>Phone:</source>
             </trans-unit>
