--- Resources/Private/Language/nl.locallang.xlf
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8" standalone="yes" ?>
-<xliff version="1.0">
-	<file source-language="en" target-language="nl" datatype="plaintext" original="messages" date="2014-03-18T11:48:24Z" product-name="pxa_dealers">
-		<header/>
-		<body>
-
-			<trans-unit id="dealers_found">
-                <source>dealers found</source>
-                <target>gevonden winkels</target>
-            </trans-unit>
-
-			<trans-unit id="error.NoResult">
-				<target>Geen verkooppunten gevonden. Probeer een andere postcode of plaats.</target>
-			</trans-unit>
-
-            <trans-unit id="categoriesFilter.radio.all">
-                <source>All</source>
-                <target>Alle winkels</target>
-            </trans-unit>
-
-			<trans-unit id="country_list_select_country">
-				<source>Select country</source>
-				<target>Land selecteren >></target>
-			</trans-unit>
-
-			<trans-unit id="country_list.all_countries">
-				<source>All countries</source>
-				<target>Alle landen</target>
-			</trans-unit>
-
-			<trans-unit id="notDealersFoundFilteringMessage">
-				<source>No retailers were found. Try searching with a different zip code or remove the filtering options.</source>
-				<target>Er zijn geen winkels gevonden. Zoek op een andere postcode of verwijder de filteropties.</target>
-			</trans-unit>
-
-		</body>
-	</file>
-</xliff>
