--- Resources/Private/Language/sv.locallang.xlf
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="utf-8" standalone="yes" ?>
-<xliff version="1.0">
-	<file source-language="en" target-language="sv" datatype="plaintext" original="messages" date="2014-03-18T11:48:24Z" product-name="pxa_dealers">
-		<header/>
-		<body>			
-
-			<trans-unit id="error.NoResult">
-				<target>Hittade inga återförsäljare. Prova att söka med ett annat postnummer eller en ort.</target>
-			</trans-unit>
-			<trans-unit id="error.NotNumeric">
-				<target>Zipcode should be numeric</target>
-			</trans-unit>
-			
-			<trans-unit id="placeholder">
-				<target>Ange postnr eller en ort</target>
-			</trans-unit>
-			
-			<trans-unit id="filtrering">
-				<target>Filtrering</target>
-			</trans-unit>
-
-			<trans-unit id="closest_dealer">
-				<target>%s Återförsäljare närmast</target>
-			</trans-unit>
-
-			<trans-unit id="notDealersFoundFilteringMessage">
-				<target>Inga butiker hittades. Försök att söka med ett annan postnummer eller ta bort filtreringsalternativ.</target>
-			</trans-unit>
-
-			<trans-unit id="filtrering_notice">
-				<target>Klicka ur kryssrutorna för att filtrera ditt resultat.</target>
-			</trans-unit>
-
-			<trans-unit id="error.EmptyValue">
-				<target>Skriv in ett postnummer eller en ort</target>
-			</trans-unit>
-
-			<trans-unit id="submitvalue">
-				<target>Sök</target>
-			</trans-unit>
-
-			<trans-unit id="ajax_loaded_dealer">
-				<target>Närmaste återförsäljare</target>
-			</trans-unit>
-
-			<trans-unit id="ajax_preload">
-				<target>Söker närmaste återförsäljare</target>
-			</trans-unit>
-			<trans-unit id="dealers_found">
-                <source>dealers found</source>
-                <target>hittade butiker</target>
-            </trans-unit>
-
-			<trans-unit id="country_list_city_zip_search">
-				<source>search for city or zip...</source>
-				<target>Ange postnummer eller ort</target>
-			</trans-unit>
-
-            <trans-unit id="categoriesFilter.radio.all">
-                <source>All</source>
-                <target>Alla Butiker</target>
-            </trans-unit>
-
-			<trans-unit id="country_list_select_country">
-				<source>Select country</source>
-				<target>Välj land >></target>
-			</trans-unit>
-
-			<trans-unit id="country_list.all_countries">
-				<source>All countries</source>
-				<target>Alla länder</target>
-			</trans-unit>
-
-		</body>
-	</file>
-</xliff>
