--- Resources/Private/Layouts/DefaultAjax.html
@@ -1,1 +1,0 @@
-<f:render section="main" />
