--- Resources/Private/Templates/Categories/List.html
@@ -1,63 +1,0 @@
-{namespace v=FluidTYPO3\Vhs\ViewHelpers}
-<f:layout name="Default" />
-
-This Template is responsible for creating a table of domain objects.
-
-If you modify this template, do not forget to change the overwrite settings
-in /Configuration/ExtensionBuilder/settings.yaml:
-Resources:
-Private:
-Templates:
-List.html: keep
-
-Otherwise your changes will be overwritten the next time you save the extension in the extension builder
-
-<f:section name="main">
-
-	<div class="pxa-dealers">
-
-		<ul class="categories">
-
-			<f:if condition="{settings.filterType} == 'checkbox'">
-				<f:then>
-
-					<f:for each="{categories}" as="category" key="catkey">
-
-						<v:variable.set name="isChecked" value="{v:variable.get(name: 'enabledCategories.{catkey}')}" />
-
-						<li class="category-item{f:if(condition: isChecked, then: ' selected')}" data-category-uid="{category.uid}">
-
-							<f:form.checkbox id="dealers-category-{category.uid}" property="dealer-category" value="1" checked='{isChecked}' />
-							<label for="dealers-category-{category.uid}">{category.title}</label>
-							<span class="icon"><i class="fa fa-check"></i></span>
-
-						</li>
-					</f:for>
-
-				</f:then>
-				<f:else>
-
-					<f:if condition="{settings.showAllMarkersOption}">
-						<li class="category-item category-item-radio category-item-all selected" data-category-uid="-1">
-							<a href="#" data-uid="-1"><f:translate key="categoriesFilter.radio.all" /></a>
-						</li>
-					</f:if>
-
-					<f:for each="{categoriesFilterOptions}" as="categoriesFilterOption">
-
-						<li class="category-item category-item-radio category-item-{categoriesFilterOption.uid}" data-category-uid="{categoriesFilterOption.categories}">
-
-							<a class="category-item-link category-item-link-{categoriesFilterOption.uid}" href="#" data-uid="{categoriesFilterOption.uid}">{categoriesFilterOption.name}</a>
-
-						</li>
-
-					</f:for>
-
-				</f:else>
-			</f:if>
-
-		</ul>
-
-	</div>
-
-</f:section>
