--- Resources/Private/Templates/CategoriesFilterOption/List.html
@@ -1,39 +1,0 @@
-
-<f:layout name="Default" />
-
-This Template is responsible for creating a table of domain objects.
-
-If you modify this template, do not forget to change the overwrite settings
-in /Configuration/ExtensionBuilder/settings.yaml:
-  Resources:
-    Private:
-      Templates:
-        List.html: keep
-
-Otherwise your changes will be overwritten the next time you save the extension in the extension builder
-
-<f:section name="main">
-<h1>Listing for CategoriesFilterOption</h1>
-
-<f:flashMessages />
-
-<table  class="tx_pxadealers" >
-	<tr>
-		<th><f:translate key="tx_pxadealers_domain_model_categoriesfilteroption.systemname" /></th>
-		<th><f:translate key="tx_pxadealers_domain_model_categoriesfilteroption.name" /></th>
-		<th> </th>
-		<th> </th>
-	</tr>
-
-	<f:for each="{categoriesFilterOptions}" as="categoriesFilterOption">
-		<tr>
-			<td><f:link.action action="show" arguments="{categoriesFilterOption : categoriesFilterOption}"> {categoriesFilterOption.systemname}</f:link.action></td>
-			<td><f:link.action action="show" arguments="{categoriesFilterOption : categoriesFilterOption}"> {categoriesFilterOption.name}</f:link.action></td>
-			<td><f:link.action action="edit" arguments="{categoriesFilterOption : categoriesFilterOption}">Edit</f:link.action></td>
-			<td><f:link.action action="delete" arguments="{categoriesFilterOption : categoriesFilterOption}">Delete</f:link.action></td>
-		</tr>
-	</f:for>
-</table>
-
-<f:link.action action="new">New CategoriesFilterOption</f:link.action>
-</f:section>
