--- Resources/Private/Templates/Countries/CountriesSelector.html
@@ -1,22 +1,0 @@
-<f:layout name="Default" />
-
-This Template is responsible for creating a table of domain objects.
-
-If you modify this template, do not forget to change the overwrite settings
-in /Configuration/ExtensionBuilder/settings.yaml:
-Resources:
-Private:
-Templates:
-List.html: keep
-
-Otherwise your changes will be overwritten the next time you save the extension in the extension builder
-
-<f:section name="main">
-
-	<div class="pxa-dealers">
-
-		<p>Not ready yet</p>
-
-	</div>
-
-</f:section>
