--- Resources/Private/Templates/Countries/List.html
@@ -1,57 +1,0 @@
-<f:layout name="Default" />
-
-This Template is responsible for creating a table of domain objects.
-
-If you modify this template, do not forget to change the overwrite settings
-in /Configuration/ExtensionBuilder/settings.yaml:
-  Resources:
-    Private:
-      Templates:
-        List.html: keep
-
-Otherwise your changes will be overwritten the next time you save the extension in the extension builder
-
-<f:section name="main">
-
-<div class="pxa-dealers">
-
-    <script type="text/javascript">
-		var countryStatesCollection = <f:format.raw>{countryZones}</f:format.raw>;
-		var countryNames = <f:format.raw>{countryNames}</f:format.raw>;
-	</script>
-
-	<f:if condition="{settings.dealers_countries_selector}">
-
-		<!-- Country selector -->
-
-	  	<div class="form-group">
-
-			<label for="dealer-countries"><f:translate key="country_list_select_country" /></label>
-			<f:form.select name="dealer-countries" class="form-control dealer-countries" property="dealer-country" options="{countries}" value="{selectedLanguage}"/>
-
-		</div>
-
-		<f:if condition="{settings.dealers_countries_states_selector}">
-
-			<!-- Country zone selector -->
-
-			<div class="form-group">
-
-				<label for="dealer-country-states"><f:translate key="country_list_select_zone" /></label>
-				<f:form.select name="dealer-country-states" class="form-control dealer-country-states" property="dealer-country-states" options='{0:"<f:translate key=\"country_list.all_country_zones\" />"}' value="0"/>
-			</div>
-
-		</f:if>
-
-	</f:if>
-
-	<f:if condition="{settings.dealers_countries_cityzip_search}">
-
-		<!-- City / zone search -->
-		<f:render partial="ZipCitySearch" />
-
-	</f:if>
-
-</div>
-
-</f:section>
