--- Resources/Private/Templates/Countries/StatesSelector.html
@@ -1,30 +1,0 @@
-<f:layout name="Default" />
-
-<f:section name="main">
-
-<div class="pxa-dealers">
-
-    <script type="text/javascript">
-		var countryStatesCollection = <f:format.raw>{countryZones}</f:format.raw>;
-	</script>
-
-	<!-- Country selector -->
-
-	<div class="form-group hidden">
-
-		<label for="dealer-countries"><f:translate key="country_list_select_country" /></label>
-		<f:form.select name="dealer-countries" class="form-control dealer-countries" property="dealer-country" options="{countries}" value="{selectedLanguage}"/>
-
-	</div>
-
-	<!-- Country zone selector -->
-
-	<div class="form-group">
-
-		<label for="dealer-country-states"><f:translate key="country_list_select_zone" /></label>
-		<f:form.select name="dealer-country-states" class="form-control dealer-country-states" property="dealer-country-states" options='{0:"<f:translate key=\"country_list.all_country_zones\" />"}' value="0"/>
-	</div>
-
-</div>
-
-</f:section>
