--- Resources/Private/Templates/Countries/ZipCitySearch.html
@@ -1,11 +1,0 @@
-<f:layout name="Default" />
-
-<f:section name="main">
-
-<div class="pxa-dealers">
-
-	<f:render partial="ZipCitySearch" />
-
-</div>
-
-</f:section>
