--- Resources/Private/Templates/Dealers/FindClosestAjax.html
@@ -1,77 +1,0 @@
-{namespace pxa=Pixelant\PxaDealers\ViewHelpers}
-{namespace v=FluidTYPO3\Vhs\ViewHelpers}
-
-<f:layout name="DefaultAjax" />
-This Template is responsible for creating a table of domain objects.
-
-If you modify this template, do not forget to change the overwrite settings
-in /Configuration/ExtensionBuilder/settings.yaml:
-  Resources:
-    Private:
-      Templates:
-        List.html: keep
-
-Otherwise your changes will be overwritten the next time you save the extension in the extension builder
-
-<f:section name="main">
-
-    <f:if condition="{dealers -> f:count()} > 0">
-        <script type="text/javascript">
-          <f:format.raw>{jsArray}</f:format.raw>
-          var settings = <v:format.json.encode value="{settings}" useTraversableKeys="0" preventRecursion="1" />
-          var labels = {}
-          labels['noResultsFoundLabel'] = "{f:translate(key: 'notDealersFoundFilteringMessage')}";
-          var markersArray = [];
-          var noResultsFoundLabel = "{f:translate(key: 'notDealersFoundFilteringMessage')}";
-          var countriesList = <f:format.raw>{countriesList}</f:format.raw>;
-        </script>
-
-        <div id="dealers-header-original" style="display:none !important"><f:translate key="closest_dealer" arguments="{0: '<span id=\"dealers-count\">{dealers -> f:count()}</span>'}" /></div>
-
-        <div class="pxa-dealers-list-container">
-
-          <div class="pxa-dealers-list{f:if(condition: settings.additionalListWrapperClass, then: ' {settings.additionalListWrapperClass}')}">
-            
-            <f:for each="{dealers}" as="dealer" iteration="itemIterationDealers">
-
-              <f:if condition="{settings.newRow}">              
-                    
-                <f:if condition="<pxa:isNewRow current=\"{itemIterationDealers.index}\" needed=\"{settings.newRow}\"/>">
-                  </div>
-                  <div class="pxa-dealers-list{f:if(condition: settings.additionalListWrapperClass, then: ' {settings.additionalListWrapperClass}')}">                   
-                </f:if>
-                  
-              </f:if>
-
-                <div  id="dealer-{itemIterationDealers.index}" class="dealer-item{f:if(condition: settings.elementAdditionalClass, then: ' {settings.elementAdditionalClass}')}" data-uid="{dealer.uid}" data-categories="{pxa:implodeCategories(categories: '{dealer.categories}')}">           
-
-                <b class="dealer-name">{dealer.name}</b>
-                <p class="dealer-info">
-                  <span class="dealer-address">{dealer.adrress}</span><br>
-                  <span class="dealer-zipcode">{dealer.zipcode}</span><br>
-                  <span class="dealer-address">{dealer.city}</span><br>
-                  <f:if condition="{dealer.website}">                    
-                      <span class="dealer-website"><f:link.external uri="{dealer.website}" target="_blank">{dealer.website}</f:link.external></span><br>
-                  </f:if>
-                  <f:if condition="{dealer.buyItNow}">                    
-                      <span class="dealer-buyItNow"><f:link.external uri="{dealer.buyItNow}" target="_blank"><f:translate key="dealers_results_buy_it_now" /></f:link.external></span><br>
-                  </f:if>
-                  <f:if condition="{dealer.telephone}">                    
-                      <span class="dealer-telephone"><a href="tel:<pxa:dealers phone='{dealer.telephone}'/>">Tel:&nbsp;{dealer.telephone}</a></span><br>
-                  </f:if> 
-                  <f:if condition="{dealer.fax}">
-                      <span class="dealer-fax"><a href="tel:<pxa:dealers phone='{dealer.fax}'/>">Fax:&nbsp;{dealer.fax}</a></span><br>
-                  </f:if>              
-                  <f:if condition="{dealer.email}">                    
-                      <span class="dealer-email"><f:link.email email="{dealer.email}" /></span><br>
-                  </f:if>
-                </p>
-              </div>
-              
-            </f:for>
-
-          </div>
-        </div>
-    </f:if>
-
-</f:section>
