--- Resources/Private/Templates/Dealers/SearchForm.html
@@ -1,24 +1,0 @@
-<f:layout name="Default" />
-
-This Template is responsible for creating a table of domain objects.
-
-If you modify this template, do not forget to change the overwrite settings
-in /Configuration/ExtensionBuilder/settings.yaml:
-  Resources:
-    Private:
-      Templates:
-        List.html: keep
-
-Otherwise your changes will be overwritten the next time you save the extension in the extension builder
-
-<f:section name="main">
-
-<div class="pxa-dealers">
-
-    <f:form action="searchResults" name="searchDealers" pageUid="{settings.searchResultPage}" pluginName="pxadealerssearchresults">
-    	<f:form.textfield name="searchValue" additionalAttributes="{placeholder: '{f:translate(key: \'country_list_city_zip_search\')}'}" />
-    	<f:form.submit name="submit" class="btn btn-primary btn-arrow" value="{f:translate(key: 'submitvalue')}" />
-    </f:form>
-
-</div>
-</f:section>
