--- Resources/Private/Templates/Dealers/SearchResults.html
@@ -1,107 +1,0 @@
-{namespace pxa=Pixelant\PxaDealers\ViewHelpers}
-{namespace v=FluidTYPO3\Vhs\ViewHelpers}
-
-<f:layout name="Default" />
-
-<f:section name="main">
-  <div class="pxa-dealers-list-wrapper{f:if(condition: settings.additionalWrapperClass, then: ' {settings.additionalWrapperClass}')}">
-
-    <f:if condition="{dealers -> f:count()} > 0">
-      <f:then>
-
-        <f:if condition="{settings.showMap}">
-          <div id="pxa-dealers-map" style="height: {f:if(condition: settings.mapHeight, then: settings.mapHeight , else: '550')}px"></div>
-        </f:if>
-
-        <script type="text/javascript">
-          var dealers = <f:format.raw>{dealersJson}</f:format.raw>
-          var pxa_dealers_enabled = 1;
-          var initialSearchValue = "<f:format.raw>{searchValue}</f:format.raw>";
-          var settings = {v:format.json.encode(value:"{settings}", useTraversableKeys:0, preventRecursion:1)}
-          var countriesList = <f:format.raw>{countriesList}</f:format.raw>
-          var labels = [];
-          labels['country_list.all_country_zones'] = '<f:translate key="country_list.all_country_zones" />';
-          labels['dealers_found'] = '<f:translate key="dealers_found" />';
-          labels['notDealersFoundFilteringMessage'] = '<f:translate key="notDealersFoundFilteringMessage" />';
-          labels['infoWindowPhone'] = '<f:translate key="infoWindowPhone" />';
-        </script>
-        <div id="dealers-header-original" style="display:none !important"><f:translate key="closest_dealer" arguments="{0: '<span id=\"dealers-count\">{dealers -> f:count()}</span>'}" /></div>
-        <h2 class="dealers-header" id="dealers-header">{dealers -> f:count()} <f:translate key="dealers_found" /></h2>
-
-        <div class="pxa-dealers-list-container{f:if(condition: '{settings.layoutType} == "alphabet"', then: ' alphabet')}"
-        id="pxa-dealers-list-container">
-
-        <f:for each="{dealers}" as="dealer" iteration="itemIterationDealers">
-
-          <div  id="dealer-{itemIterationDealers.index}" class="dealer-item isotope-item{f:if(condition: settings.elementAdditionalClass, then: ' {settings.elementAdditionalClass}')}"
-                data-name="{dealer.name}"
-                data-uid="{dealer.uid}"
-                data-categories="{pxa:implodeCategories(categories: '{dealer.categories}')}"
-                data-marker-type="default">
-
-            <div class="dealer-header">
-              <b class="dealer-name">{dealer.name}</b>
-            </div>
-            <div id="dealer-body-{itemIterationDealers.index}" class="dealer-body">
-              <f:if condition="{settings.showLogo}">
-                <div class="dealer-logo">
-                  <f:if condition="{dealer.logo}">
-                    <f:then>
-                      <f:image src="uploads/tx_pxadealers/{dealer.logo}" alt="{dealer.name}" width="{settings.dealerLogoWidth}" height="{settings.dealerLogoHeight}" />
-                    </f:then>
-                    <f:else>
-                      <f:image src="{settings.noImagePath}" alt="{dealer.name}" width="{settings.dealerLogoWidth}" height="{settings.dealerLogoHeight}" />
-                    </f:else>
-                  </f:if>
-                </div>
-              </f:if>
-              <p class="dealer-info">
-                <span class="dealer-address">{dealer.adrress}</span><br>
-                <span class="dealer-zipcode">{dealer.zipcode}</span><br>
-                <span class="dealer-address">{dealer.city}</span><br>
-                <f:if condition="{dealer.website}">
-                  <span class="dealer-website"><f:link.external uri="{dealer.website}" target="_blank">{dealer.website}</f:link.external></span><br>
-                </f:if>
-                <f:if condition="{dealer.buyItNow}">
-                  <span class="dealer-buyItNow"><f:link.external uri="{dealer.buyItNow}" target="_blank"><f:translate key="dealers_results_buy_it_now" /></f:link.external></span><br>
-                </f:if>
-                <f:if condition="{dealer.telephone}">
-                  <span class="dealer-telephone"><a href="tel:<pxa:dealers phone='{dealer.telephone}'/>"><f:translate key="infoWindowPhone" />&nbsp;{dealer.telephone}</a></span><br>
-                </f:if>
-                <f:if condition="{dealer.fax}">
-                  <span class="dealer-fax"><a href="tel:<pxa:dealers phone='{dealer.fax}'/>">Fax:&nbsp;{dealer.fax}</a></span><br>
-                </f:if>
-                <f:if condition="{dealer.email}">
-                  <span class="dealer-email"><f:link.email email="{dealer.email}" /></span><br>
-                </f:if>
-              </p>
-            </div>
-          </div>
-
-        </f:for>
-
-  </div>
-  </f:then>
-  <f:else>
-    <div class="alert alert-error">
-      <button type="button" class="close" data-dismiss="alert">×</button>
-      <f:translate key="error.NoResult" />
-    </div>
-
-    <f:if condition="{settings.showMap}">
-      <div id="pxa-dealers-map" style="height: {f:if(condition: settings.mapHeight, then: settings.mapHeight , else: '550')}px"></div>
-    </f:if>
-
-    <div id="ajax-load-dealer">
-      <h2 class="dealers-header pre-loader">
-        <f:translate key="ajax_preload" />
-        <f:image src="EXT:pxa_dealers/Resources/Public/Icons/ajax_loader_blue.gif" alt="ajax-loader" />
-      </h2>
-      <h2 class="dealers-header after-load"><f:translate key="ajax_loaded_dealer" /></h2>
-    </div>
-  </f:else>
-  </f:if>
-
-  </div>
-
-</f:section>
