+++ Resources/Private/Templates/Filter/CategoriesCollectionFilter.html
@@ -1,0 +1,17 @@
+<div xmlns="http://www.w3.org/1999/xhtml" lang="en"
+     xmlns:f="http://typo3.org/ns/TYPO3/Fluid/ViewHelpers">
+
+    <f:layout name="Default" />
+
+    <f:section name="main">
+        <f:render partial="Javascript/JsFilterCheckboxSettings" section="main" arguments="{uid: uid}"/>
+
+        <div id="pxa-dealer-map-filter-{uid}" class="pxa-dealers-categories">
+            <f:for each="{categoriesCollections}" as="categoriesCollection">
+                <div class="checkbox">
+                    <label><input type="checkbox" value="{f:if(condition: iteration.isLast, then: '{cat.uid}', else: '{cat.uid},') -> f:for(each: categoriesCollection.categories, as: 'cat', iteration: 'iteration')}"> {categoriesCollection.name}</label>
+                </div>
+            </f:for>
+        </div>
+    </f:section>
+</div>
